requirements_dev.txt
cython
sphinx
sphinx_rtd_theme
sphinx-gallery
sphinxcontrib-apidoc
pillow
pytest
pytest-cov
pylint
sympy
wheel
twine
vtk
pyvista>=0.20.1
numpydoc
nbsphinx
coverage
codecov
omf
matplotlib
pymatsolver>=0.1.2
